server/requirements.txt
anyio==3.7.0
asgiref==3.7.2
autopep8==2.0.2
certifi==2023.5.7
Django==4.2.3
django-autocomplete-light==3.9.7
django-filter==23.2
django-markdownx==4.0.2
django-model-utils==4.3.1
djangorestframework==3.14.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
Markdown==3.4.3
phonenumbers==8.13.15
Pillow==10.0.0
pycodestyle==2.10.0
python-markdown-math==0.8
python-telegram-bot==20.3
pytz==2023.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
